Major Changes
Completely redid the simulation framework so that it would use CRTP to
help remove a lot of costly run time polymorphism. To roughly test
run-time costs, I wrote a simple Timer class.

I have also completely re-coded the Neural Network to be based on a
graph-like implementation that I hope to extend so that it can be used
in a recursive neural network, as well as the traditional feed forward
neural network.

I have yet to implement the training algorithms for the neural network,
but the feed forward computation has been checked and validated.