Machine learning Removal and Additions, plus more
Removed old machine learning stuff for ANN, replaced with a graph
formulation of an ANN so that it can later be extended to an RNN.

Also added in files for PDE solution methods.

Also added in a file object that will handle freeing a FILE pointer,
but still allow me to use C based file IO methods (which I prefer).