Matrix Updates
Setup the matrices to be built into a unique type using CRTP so I can
add methods to each of the matrices that may tailor performance to them.

Took a bit of work, but found the problem I was having with some
implicitly deleted copy assignment operator was related to not
specifying a specific operator= for setting equal to the same time
(which is ironic because I had an operator= template.