Optimization Clean-Up and Building
Removed training algorithms built for the neural network because the
end goal is to use the optimization code to train neural networks.

Additionally, configured a template based approach to piece together
optimization algorithms based on gradients. Should essentially be
working by having an optimization class have parameters for the cost
function and the step function, which will be based on the function as
well. So far so good.