Adding in new Filters, working on ANN
Spent some time trying to build some new filters, one being the ADELINE
which is an adaptive filter based on a neural network. This is still a
work in progress because, while I thought I verified my neural network
before, I still have some work to do with the ANN itself.

I put an implementation in place for a Recursive Least Square
estimator. Application might be to estimate position but use the
hopefully better estimates to fit a parametric velocity and
acceleration time sequence. Still a work in progress.