Validated Neural Network and Fixed Bug in ADELINE Filter
The main changes is a validation of the neural network and fixing a bug
in the ADELINE filter. One other note is the ADELINE filter was
modified to actually have a hidden layer with sigmoid activation
functions. In an example problem I ran in TestADELINE_Filter1.hpp, this
change made the filter perform better than in the case where ADELINE is
learning an Affine transformation.