Major changes to ANN, Modified PSO, Added to Gradient-based Algorithms, Added Tests
Fixed major bugs in ANN and actually re-architected it to represent
connections between layers as matrices, which seems to be fairly
standard. This has made much of the algorithmic nature of the ANN much
simpler and easier to understand.

With respect to the ANN, also added a Leaky ReLU activation function
that is better suited to deep learning problems. Also built a few test
functions to try out the ANN. Two of them are more to see it run on a
classification problem and a dimensionality reduction problem. The
third test code is checking that the weight gradient it outputs matches
what is expected (so a sanity check). The truth result is one based on
an analytical representation derived by hand.

For the PSO, found a bug in the RandomInit class that was making it
fail to initialize the swarm properly. They ended up all only differing
in the first dimension of their vectors. This fix improved the
algorithm’s performance. Still looking to see if anything could be
improved. Found one new add on was an item called Inertia, which can be
used to influence how much the prior velocity propagates into the new
one via the update formula.

As for the gradient-based algorithms, implemented a stepsize algorithm
based on a Hessian-Free approximation for a local quadratic model
around a given point. Seems to work well often, but not sure if it’s
perfectly implemented yet because it has blown up for some ANN problems
I used it on. Can’t tell if it blew up, though, because the Hessian was
technically not positive definite at those locations.

Anyway, those are the main changes made thus far.