Working on UKF, PSO, Linear Algebra Stuff
For UKF, trying to make the filter based on a cholesky factorization
and rank one updates.

For PSO, trying to add in capability to select various neighborhoods.

For linear algebra stuff, added in rank one updates/downdates, and am
looking into what appears to be round off errors for LDL and possibly
even cholesky factorizations.