Fixed LDL Decomposition
Got the LDL decompositional algorithm working fine after referencing a
separate version of the algorithm via Matrix Computations 4th Edition.

With this new LDL, got the UKF working the way I would hope. What’s
different is what happens in the update equation:
Pxx = Pxx - K*Pzz*K’

Assuming M = K*Pzz*K’, and Pxx is decomposed into cholesky factors, I
wanted the cholesky factor of Pxx to be modified by a sequence of rank
one downdates based on M. To do this rank one downdates, instead of
making M into a cholesky factorization, I turn it into an LDL
factorization. The reasoning is because if any term on the LDL
decomposition is negative, then I am able to set it to zero to ensure M
is positive definite (which I believe it should be, though numerically
it might lose that property). I also use the LDL because if any terms
on the diagonal are zero, then their corresponding columns of L don’t
need to be incorporated into the downdate sequence, thus making the
algorithm slightly faster.