Fixed some Linear Algebra stuff, did some filter work
Fixed some template mistakes in the MatHashes.* that were resulting in
undefined behavior, which happened to not do what I wanted to when I
put compiler optimizations on. The optimized code now matches with the
debug compiled code.

Modified Unscented Kalman Filter (UKF) such that it was cleaned up and
runs with a covariance matrix Cholesky factorization and LDL
factorization rank-one downdate for the update formula of the
covariance matrix which helps remove potential for becoming negative
definite. Still need to investigate because I saw in the
TestExample1.hpp, setting the noise to 8e-5 (i believe) results in NaN.

Cleaned up the Neural Filter and removed some unnecessary code, like
testing algorithms for an adaptive step size based on variation in the
time series value. Will have to consider architecting this filter such
that a user might be able to specify their own stepsize function..
maybe in the same way I specify stepsize functions for the gradient
based optimization codes (?).

Removed some files that were not really needed.

Added linear kalman filter code semi-based on the UKF code. removed
lots of stuff that wasn’t needed and made the necessary modifications
that would allow the linear kalman filter code to be fast. Only thing
that might be improved is potentially looking into doing a cholesky
factorization thing for the Linear Kalman Filter just like in the UKF.