Cleaned up some test code, tweaked Fraction, started Cross Validation framework
Cleaned up ANN and Filtering test code a little bit. Changed Fraction
so that it aims to get within 1e-4 of some double value (may change
this back). Noticed the fraction class struggled to approximate pi.
Will have to understand why.

Started the code structure for a Cross Validation framework. It’s
mostly done, but needs to be tested with a data set and model class.
The point of the test will also be to figure out what methods a person
would have to build and how they should perform, given that the data
set and model are template parameters to the CV class.