Completed a Cross Validation framework with samples, added test automation
Finished and tested the cross validation framework against a sample
problem fitting quadratic data to linear, quadratic, and cubic models
to see which model is the best.

In the example, the framework shows quadratic being best and cubic
being next best, though improving as the data size increases. This
shows the framework is doing as expected.