Complete most of new Neural Network implementation
Completed most of the new Neural Network implementation and have shown
things to be working properly. Still need to implement the back
propagation algorithm.

Added in a class that represents the prototype design pattern, passing
references to functors to the nodes of the ANN. This saves memory and
is efficient. May check performance comparison to see how it differs
from using normal C function pointers.

Got the prototype class integrated into the net layer and whatnot. Have
tested the ANN forward propagation and it appears to be working
properly.